Tweak Paragraph story
